BAEL-3141: Created new module for 'Converting Java String to Double' and other newly created string articles.
