+++ java-strings-3/src/main/java/com/baeldung/string/README.md
@@ -1,0 +1,3 @@
+This file exists to ensure this empty directory is committed in Git.
+
+Please remove this file when this directory is populated.
